chore: fixing linting and moving functions out of HabitEditor
