+++ src/Components/HabitEditor/HabitEditor.tsx
@@ -5,17 +5,14 @@ import React, { useCallback, useContext, useEffect, useRef, useState } from 'rea
 import {
     View,
     Text,
-    ScrollView,
-    Dimensions,
-    StyleSheet,
     TouchableOpacity,
     Keyboard,
     EmitterSubscription,
     InteractionManager
 } from 'react-native';
-import { TextInput, TouchableWithoutFeedback } from 'react-native-gesture-handler';
+import { TextInput } from 'react-native-gesture-handler';
 import Animated from 'react-native-reanimated';
-import { ColourPicker, randomGradient } from 'Components/ColourPicker';
+import { ColourPicker } from 'Components/ColourPicker';
 import Icon, { IconProps } from 'Components/Icon';
 import { GradientColours, GreyColours } from 'Styles/Colours';
 import BottomSheet from 'reanimated-bottom-sheet';
@@ -32,11 +29,13 @@ import { HabitType } from 'Components/Habit';
 import { getRandomBytes } from 'expo-random';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
 import { Card } from 'Components/Card';
-import { TimePicker } from 'Components/TimePicker';
 import Toast, { BaseToastProps } from 'react-native-toast-message';
 import { AppContext } from 'Context/AppContext';
 import { impactAsync, ImpactFeedbackStyle, notificationAsync, NotificationFeedbackType } from 'expo-haptics';
 import { EditNavProps } from 'Screens/Edit';
+import { globalStyles } from './HabitEditor.styles';
+import { randomGradient } from 'Helpers';
+import { ShadowModal, IconModal, HeaderModal, TimeModal } from './molecules';
 
 interface EditProps {
 	navigation: EditNavProps;
@@ -475,476 +474,5 @@ const HabitEdtor: React.FC<EditProps> = ({ navigation, habit }) => {
         </React.Fragment>
     );
 };
-
-interface ShadowModalProps {
-	shadow: Animated.Value<number>;
-}
-
-const ShadowModal = ({ shadow }: ShadowModalProps) => {
-    const animatedShadowOpacity = Animated.interpolate(shadow, {
-        inputRange: [0, 1],
-        outputRange: [0.5, 0]
-    });
-
-    return (
-        <Animated.View
-            pointerEvents='none'
-            style={[
-			  modalStyles.shadowContainer,
-			  {
-			    opacity: animatedShadowOpacity
-			  }
-            ]}
-        />
-    );
-};
-
-interface TimeModalProps {
-	minutes: number;
-	hours: number;
-	handleTimeType: (values: { hours: number; minutes: number }) => void;
-}
-
-const TimeModal = ({ minutes, hours, handleTimeType }: TimeModalProps) => {
-    const { colors } = useTheme();
-
-    return (
-        <View
-            style={{
-			  backgroundColor: colors.card,
-			  padding: 16,
-			  height: '100%'
-            }}
-        >
-            <TimePicker value={{ hours, minutes }} hoursUnit='hr' minutesUnit='m' onChange={handleTimeType} />
-        </View>
-    );
-};
-
-interface HeaderModalProps {
-	sheetRef: React.RefObject<BottomSheet>;
-}
-
-interface IconGroup {
-	label: string;
-	icons: Partial<IconProps>[];
-}
-
-const IconOptions: IconGroup[] = [
-    {
-        label: 'Sport',
-        icons: [
-            { family: 'materialcommunity', name: 'soccer' },
-            { family: 'materialcommunity', name: 'volleyball' },
-            { family: 'materialcommunity', name: 'basketball' },
-            { family: 'materialcommunity', name: 'baseball' },
-            { family: 'materialcommunity', name: 'tennis-ball' },
-            { family: 'materialcommunity', name: 'billiards' },
-            { family: 'materialcommunity', name: 'bowling' },
-            { family: 'materialcommunity', name: 'bullseye-arrow' },
-            { family: 'materialcommunity', name: 'rugby' },
-            { family: 'materialcommunity', name: 'football-helmet' },
-            { family: 'materialcommunity', name: 'boxing-glove' },
-            { family: 'materialcommunity', name: 'basketball-hoop' },
-            { family: 'materialcommunity', name: 'baseball-bat' },
-            { family: 'materialcommunity', name: 'tennis' },
-            { family: 'materialcommunity', name: 'table-tennis' },
-            { family: 'materialcommunity', name: 'golf' },
-            { family: 'materialcommunity', name: 'hockey-sticks' },
-            { family: 'materialcommunity', name: 'dumbbell' },
-            { family: 'materialcommunity', name: 'weight-lifter' },
-            { family: 'materialcommunity', name: 'walk' },
-            { family: 'materialcommunity', name: 'run' },
-            { family: 'materialcommunity', name: 'ski-cross-country' },
-            { family: 'materialcommunity', name: 'ski' },
-            { family: 'materialcommunity', name: 'ski-water' },
-            { family: 'materialcommunity', name: 'swim' },
-            { family: 'materialcommunity', name: 'bike' },
-            { family: 'materialcommunity', name: 'rowing' },
-            { family: 'materialcommunity', name: 'karate' },
-            { family: 'materialcommunity', name: 'yoga' },
-            { family: 'materialcommunity', name: 'meditation' }
-        ]
-    },
-    {
-        label: 'Art',
-        icons: [
-            { family: 'materialcommunity', name: 'brush' },
-            { family: 'materialcommunity', name: 'pencil' },
-            { family: 'materialcommunity', name: 'fountain-pen-tip' },
-            { family: 'materialcommunity', name: 'palette' },
-            { family: 'materialcommunity', name: 'spray' },
-            { family: 'materialcommunity', name: 'drama-masks' },
-            { family: 'materialcommunity', name: 'piano' },
-            { family: 'materialcommunity', name: 'microphone-variant' },
-            { family: 'materialcommunity', name: 'saxophone' },
-            { family: 'materialcommunity', name: 'violin' },
-            { family: 'materialcommunity', name: 'guitar-acoustic' },
-            { family: 'materialcommunity', name: 'guitar-electric' },
-            { family: 'materialcommunity', name: 'trumpet' },
-            { family: 'materialcommunity', name: 'instrument-triangle' },
-            { family: 'materialcommunity', name: 'album' },
-            { family: 'materialcommunity', name: 'movie-roll' },
-            { family: 'materialcommunity', name: 'movie-open' },
-            { family: 'materialcommunity', name: 'camera' },
-            { family: 'materialcommunity', name: 'music' },
-            { family: 'materialcommunity', name: 'music-clef-treble' },
-            { family: 'materialcommunity', name: 'music-clef-bass' }
-        ]
-    },
-    {
-        label: 'Lifestyle',
-        icons: [
-            { family: 'materialcommunity', name: 'book-open-variant' },
-            { family: 'materialcommunity', name: 'school' },
-            { family: 'materialcommunity', name: 'brain' },
-            { family: 'materialcommunity', name: 'sleep' },
-            { family: 'materialcommunity', name: 'bed' },
-            { family: 'materialcommunity', name: 'shower' },
-            { family: 'materialcommunity', name: 'chef-hat' },
-            { family: 'materialcommunity', name: 'watering-can' },
-            { family: 'materialcommunity', name: 'hanger' },
-            { family: 'materialcommunity', name: 'hammer-wrench' },
-            { family: 'materialcommunity', name: 'broom' },
-            { family: 'materialcommunity', name: 'toothbrush' },
-            { family: 'materialcommunity', name: 'trash-can' },
-            { family: 'materialcommunity', name: 'washing-machine' },
-            { family: 'materialcommunity', name: 'dog-service' },
-            { family: 'materialcommunity', name: 'pill' },
-            { family: 'materialcommunity', name: 'needle' },
-            { family: 'materialcommunity', name: 'hospital-box' },
-            { family: 'materialcommunity', name: 'cellphone' },
-            { family: 'materialcommunity', name: 'television-classic' },
-            { family: 'materialcommunity', name: 'laptop' },
-            { family: 'materialcommunity', name: 'desktop-classic' },
-            { family: 'materialcommunity', name: 'gamepad-variant' },
-            { family: 'materialcommunity', name: 'headphones' },
-            { family: 'materialcommunity', name: 'newspaper' },
-            { family: 'materialcommunity', name: 'cash-multiple' },
-            { family: 'materialcommunity', name: 'tent' },
-            { family: 'materialcommunity', name: 'image-filter-hdr' }
-        ]
-    },
-    {
-        label: 'Emoticons',
-        icons: [
-            { family: 'materialcommunity', name: 'emoticon' },
-            { family: 'materialcommunity', name: 'emoticon-angry' },
-            { family: 'materialcommunity', name: 'emoticon-confused' },
-            { family: 'materialcommunity', name: 'emoticon-cool' },
-            { family: 'materialcommunity', name: 'emoticon-cry' },
-            { family: 'materialcommunity', name: 'emoticon-excited' },
-            { family: 'materialcommunity', name: 'emoticon-frown' },
-            { family: 'materialcommunity', name: 'emoticon-happy' },
-            { family: 'materialcommunity', name: 'emoticon-kiss' },
-            { family: 'materialcommunity', name: 'emoticon-lol' },
-            { family: 'materialcommunity', name: 'emoticon-neutral' },
-            { family: 'materialcommunity', name: 'emoticon-sad' },
-            { family: 'materialcommunity', name: 'emoticon-tongue' },
-            { family: 'materialcommunity', name: 'emoticon-wink' },
-            { family: 'materialcommunity', name: 'emoticon-poop' }
-        ]
-    },
-    {
-        label: 'Food',
-        icons: [
-            { family: 'materialcommunity', name: 'food-apple' },
-            { family: 'materialcommunity', name: 'fruit-cherries' },
-            { family: 'materialcommunity', name: 'fruit-grapes' },
-            { family: 'materialcommunity', name: 'fruit-pineapple' },
-            { family: 'materialcommunity', name: 'fruit-watermelon' },
-            { family: 'materialcommunity', name: 'food-croissant' },
-            { family: 'materialcommunity', name: 'baguette' },
-            { family: 'materialcommunity', name: 'cheese' },
-            { family: 'materialcommunity', name: 'food-drumstick' },
-            { family: 'materialcommunity', name: 'food-steak' },
-            { family: 'materialcommunity', name: 'food' },
-            { family: 'materialcommunity', name: 'food-variant' },
-            { family: 'materialcommunity', name: 'noodles' },
-            { family: 'materialcommunity', name: 'pasta' },
-            { family: 'materialcommunity', name: 'hamburger' },
-            { family: 'materialcommunity', name: 'cupcake' },
-            { family: 'materialcommunity', name: 'coffee' },
-            { family: 'materialcommunity', name: 'beer' }
-        ]
-    },
-    {
-        label: 'Social',
-        icons: [
-            { family: 'materialcommunity', name: 'facebook' },
-            { family: 'materialcommunity', name: 'facebook-messenger' },
-            { family: 'materialcommunity', name: 'instagram' },
-            { family: 'materialcommunity', name: 'youtube' },
-            { family: 'materialcommunity', name: 'linkedin' },
-            { family: 'materialcommunity', name: 'twitch' },
-            { family: 'materialcommunity', name: 'twitter' },
-            { family: 'materialcommunity', name: 'reddit' },
-            { family: 'materialcommunity', name: 'steam' },
-            { family: 'materialcommunity', name: 'microsoft-xbox' },
-            { family: 'materialcommunity', name: 'sony-playstation' },
-            { family: 'materialcommunity', name: 'snapchat' }
-        ]
-    },
-    {
-        label: 'Miscellaneous',
-        icons: [
-            { family: 'materialcommunity', name: 'one-up' },
-            { family: 'materialcommunity', name: 'atom' },
-            { family: 'materialcommunity', name: 'rocket' },
-            { family: 'materialcommunity', name: 'atom-variant' },
-            { family: 'materialcommunity', name: 'medal' },
-            { family: 'materialcommunity', name: 'trophy' },
-            { family: 'materialcommunity', name: 'moon-waning-crescent' },
-            { family: 'materialcommunity', name: 'star' },
-            { family: 'materialcommunity', name: 'death-star-variant' },
-            { family: 'materialcommunity', name: 'white-balance-sunny' },
-            { family: 'materialcommunity', name: 'heart' },
-            { family: 'materialcommunity', name: 'hand-heart' },
-            { family: 'materialcommunity', name: 'account-heart' },
-            { family: 'materialcommunity', name: 'cards' },
-            { family: 'materialcommunity', name: 'cards-club' },
-            { family: 'materialcommunity', name: 'cards-diamond' },
-            { family: 'materialcommunity', name: 'cards-spade' }
-        ]
-    }
-];
-
-interface IconModalProps {
-	setIcon: React.Dispatch<React.SetStateAction<Partial<IconProps>>>;
-	closeSheet: () => void;
-}
-
-const IconModal = ({ setIcon, closeSheet }: IconModalProps) => {
-    const { colors } = useTheme();
-    const iconWidth = Dimensions.get('screen').width / 6.6;
-
-    const handlePress = (icon: Partial<IconProps>) => {
-        setIcon(icon);
-        closeSheet();
-    };
-    return (
-        <ScrollView
-            style={{
-			  backgroundColor: colors.card,
-			  padding: 16,
-
-			  height: '100%'
-            }}
-            showsVerticalScrollIndicator={false}
-        >
-            <View style={{ paddingBottom: 50 }}>
-                {IconOptions.map((group, index) => (
-                    <View key={group.label}>
-                        <View
-                            style={{
-							  flexDirection: 'row',
-							  marginTop: index === 0 ? 0 : 20,
-							  marginBottom: 5,
-							  marginLeft: 10
-                            }}
-                        >
-                            <Text
-                                style={{
-								  padding: 10,
-								  backgroundColor: colors.background,
-								  color: colors.text,
-								  fontSize: 16,
-								  fontFamily: 'Montserrat_700Bold',
-								  borderRadius: 10,
-								  overflow: 'hidden'
-                                }}
-                            >
-                                {group.label}
-                            </Text>
-                        </View>
-
-                        <View
-                            style={{
-							  width: '100%',
-							  display: 'flex',
-							  flexDirection: 'row',
-							  flexWrap: 'wrap'
-                            }}
-                        >
-                            {group.icons.map((icon, index) => (
-                                <TouchableOpacity
-                                    key={index + icon.name}
-                                    onPress={() => handlePress(icon)}
-                                    style={{
-									  padding: 10,
-									  width: iconWidth,
-									  display: 'flex',
-									  alignItems: 'center'
-                                    }}
-                                >
-                                    <Icon family={icon.family} name={icon.name} colour={colors.text} size={30} />
-                                </TouchableOpacity>
-                            ))}
-                        </View>
-                    </View>
-                ))}
-            </View>
-        </ScrollView>
-    );
-};
-
-interface HeaderModalProps {
-	sheetRef: React.RefObject<BottomSheet>;
-	height: number;
-}
-
-const HeaderModal = ({ sheetRef, height }: HeaderModalProps) => {
-    const { colors } = useTheme();
-    const closeSheet = () => {
-        sheetRef.current && sheetRef.current.snapTo(1);
-    };
-
-    const { header } = modalStyles;
-    const headerBackground = { backgroundColor: colors.card };
-    const headerStyles = StyleSheet.flatten([header, headerBackground]);
-
-    return (
-        <TouchableWithoutFeedback style={[modalStyles.headerTouchable, { height: height }]} onPress={closeSheet}>
-            <View style={headerStyles}>
-                <View style={modalStyles.panelHeader}>
-                    <View style={modalStyles.panelHandle} />
-                </View>
-            </View>
-        </TouchableWithoutFeedback>
-    );
-};
-
-const modalStyles = StyleSheet.create({
-    header: {
-        borderTopLeftRadius: 20,
-        borderTopRightRadius: 20,
-        height: 40,
-        paddingTop: 10,
-        shadowColor: '#000000'
-    },
-    headerTouchable: {
-        display: 'flex',
-        justifyContent: 'flex-end'
-    },
-    panelHandle: {
-        backgroundColor: '#00000040',
-        borderRadius: 4,
-        height: 8,
-        marginBottom: 10,
-        width: 40
-    },
-    panelHeader: {
-        alignItems: 'center'
-    },
-    shadowContainer: {
-        ...StyleSheet.absoluteFillObject,
-        backgroundColor: '#000',
-        zIndex: 2
-    }
-});
-
-const globalStyles = StyleSheet.create({
-    cardText: {
-        fontFamily: 'Montserrat_600SemiBold',
-        fontSize: 18
-    },
-    count: {
-        alignItems: 'center',
-        borderRadius: 5,
-        height: 45,
-        justifyContent: 'center',
-        overflow: 'hidden',
-        width: 45
-    },
-
-    gradient: {
-        bottom: 0,
-        left: 0,
-        position: 'absolute',
-        right: 0,
-        top: 0
-    },
-    type: {
-        alignItems: 'center',
-        borderRadius: 5,
-        display: 'flex',
-        height: 45,
-        justifyContent: 'center',
-        overflow: 'hidden',
-        width: 45
-    }
-});
-
-// icons: [
-// 	{ family: 'feather', name: 'book' },
-// 	{ family: 'feather', name: 'award' },
-// 	{ family: 'feather', name: 'book-open' },
-// 	{ family: 'feather', name: 'bookmark' },
-// 	{ family: 'feather', name: 'briefcase' },
-// 	{ family: 'feather', name: 'camera' },
-// 	{ family: 'feather', name: 'compass' },
-// 	{ family: 'feather', name: 'film' },
-// 	{ family: 'feather', name: 'gift' },
-// 	{ family: 'feather', name: 'headphones' },
-// 	{ family: 'feather', name: 'heart' },
-// 	{ family: 'feather', name: 'moon' },
-// 	{ family: 'feather', name: 'sun' },
-// 	{ family: 'feather', name: 'monitor' },
-// 	{ family: 'feather', name: 'music' },
-// 	{ family: 'feather', name: 'pen-tool' },
-// 	{ family: 'feather', name: 'smile' },
-// 	{ family: 'feather', name: 'tablet' },
-// 	{ family: 'feather', name: 'database' },
-// 	{ family: 'feather', name: 'server' },
-// 	{ family: 'feather', name: 'star' },
-// 	{ family: 'feather', name: 'thumbs-up' },
-// 	{ family: 'feather', name: 'phone' },
-// 	{ family: 'feather', name: 'tool' },
-// 	{ family: 'feather', name: 'tv' },
-// 	{ family: 'feather', name: 'twitch' },
-// 	{ family: 'feather', name: 'twitter' },
-// 	{ family: 'feather', name: 'instagram' },
-// 	{ family: 'feather', name: 'facebook' },
-// 	{ family: 'feather', name: 'linkedin' },
-// 	{ family: 'feather', name: 'youtube' },
-// 	{ family: 'feather', name: 'video' },
-// 	{ family: 'materialcommunity', name: 'baguette' },
-// 	{ family: 'materialcommunity', name: 'baseball' },
-// 	{ family: 'materialcommunity', name: 'baseball-bat' },
-// 	{ family: 'materialcommunity', name: 'basketball' },
-// 	{ family: 'materialcommunity', name: 'basketball-hoop' },
-// 	{ family: 'materialcommunity', name: 'bed' },
-// 	{ family: 'materialcommunity', name: 'beer' },
-// 	{ family: 'materialcommunity', name: 'bike' },
-// 	{ family: 'materialcommunity', name: 'billiards' },
-// 	{ family: 'materialcommunity', name: 'bowling' },
-// 	{ family: 'materialcommunity', name: 'brain' },
-// 	{ family: 'materialcommunity', name: 'calculator' },
-// 	{ family: 'materialcommunity', name: 'calendar' },
-// 	{ family: 'materialcommunity', name: 'calendar-check' },
-// 	{ family: 'materialcommunity', name: 'calendar-heart' },
-// 	{ family: 'materialcommunity', name: 'camera' },
-// 	{ family: 'materialcommunity', name: 'car-sports' },
-// 	{ family: 'materialcommunity', name: 'cart' },
-// 	{ family: 'materialcommunity', name: 'cards' },
-// 	{ family: 'materialcommunity', name: 'cards-club' },
-// 	{ family: 'materialcommunity', name: 'cards-didmond' },
-// 	{ family: 'materialcommunity', name: 'cards-heart' },
-// 	{ family: 'materialcommunity', name: 'cards-spade' },
-// 	{ family: 'materialcommunity', name: 'charity' },
-// 	{ family: 'materialcommunity', name: 'chart-areaspline-variant' },
-// 	{ family: 'materialcommunity', name: 'chart-bar' },
-// 	{ family: 'materialcommunity', name: 'chart-pie' },
-// 	{ family: 'materialcommunity', name: 'chef-hat' },
-// 	{ family: 'materialcommunity', name: 'cupcake' },
-
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// 	// { family: 'materialcommunity', name: 'emoticon-' },
-// ],
 
 export default HabitEdtor;
