+++ src/Components/HabitEditor/molecules/index.ts
@@ -1,0 +1,4 @@
+export * from './HeaderModal';
+export * from './IconModal';
+export * from './ShadowModal';
+export * from './TimeModal';
