+++ src/Components/TimePicker/index.ts
@@ -1,1 +1,2 @@
+export { default } from './TimePicker';
 export * from './TimePicker';
