+++ src/Screens/Awards/AwardsScreen.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 
-export default function AwardsScreen () {
+const AwardsScreen: React.FC = () => {
     return (
         <View>
             <Text>Awards</Text>
         </View>
     );
-}
+};
+
+export default AwardsScreen;
